separate fetcher of api handlers
