+++ pages/event/[eventId].tsx
@@ -1,7 +1,7 @@
 import { NextPage } from "next";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
-import { getEventById } from "../../lib/fetcher";
+import { getEventById } from "../../lib/api-handlers";
 import { Spinner } from "../../stories/Spinner/Spinner";
 import { EventData } from "../../types/events-types";
