+++ pages/event/list.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { getAllEvents } from "../../lib/fetcher";
+import { getAllEvents } from "../../lib/api-handlers";
 import { Spinner } from "../../stories/Spinner/Spinner";
 import { EventData } from "../../types/events-types";
